add gpio write casts

+++ gpio.c
@@ -30,16 +30,16 @@ void gpio_write(GPIO_TypeDef *port, uint8_t pin, bool value) {
         return;
     }
     if (value)
-        port->ODR |= (1U << pin);
+        port->ODR |= (1U << (uint32_t)pin);
     else
-        port->ODR &= ~(1U << pin);
+        port->ODR &= ~(1U << (uint32_t)pin);
 }
 
 void gpio_toggle(GPIO_TypeDef *port, uint8_t pin) {
     if (!gpio_is_valid(port, pin)) {
         return;
     }
-    port->ODR ^= (1U << pin);
+    port->ODR ^= (1U << (uint32_t)pin);
 }
 
 bool gpio_read(GPIO_TypeDef *port, uint8_t pin) {
